Restructure packages to group classes by layer within feature
