+++ src/main/java/dev/austinbarnes/retailinventorymanagement/location/entity/WarehouseLocation.java
@@ -1,4 +1,4 @@
-package dev.austinbarnes.retailinventorymanagement.location;
+package dev.austinbarnes.retailinventorymanagement.location.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Size;
@@ -10,19 +10,20 @@ import lombok.Setter;
 import java.util.UUID;
 
 @Entity
-@Table(name = "retail_location")
+@Table(name = "warehouse")
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
-public class RetailLocation {
+public class WarehouseLocation {
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
 
-    @Column(name = "retail_location_code")
+    //TODO: Create method for auto generated codes. Use util and another table?
+    @Column(name = "warehouse_code")
     @Size(min = 6, max = 6)
-    private String retailLocationCode;
+    private String warehouseCode;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "location_id", referencedColumnName = "id")
@@ -30,4 +31,5 @@ public class RetailLocation {
 
     @Column(name = "deleted")
     private boolean deleted = false;
+
 }
